:truck: (901): move llm prompts from services to infrastructure/llm/prompts
