+++ mcr-core/mcr_meeting/app/infrastructure/llm/prompts/acronyms.py
