+++ mcr-core/mcr_meeting/app/infrastructure/llm/prompts/data/glossary.md
