+++ mcr-core/mcr_meeting/app/infrastructure/llm/prompts/participants.py
