+++ mcr-core/mcr_meeting/evaluation/acronyms/evaluation_script.py
@@ -54,7 +54,7 @@ REFERENCE_DIR = DATA_DIR / "references"
 OUTPUT_DIR = EVALUATION_DIR / "data" / "outputs" / "acronyms_outputs"
 
 # In our dataset, which is gitignored (in the data dir), we have the three following files:
-# - in_glossary.json: list of acronyms that are in the glossary markdown file mcr_meeting/app/services/correct_acronyms/data/glossary.md
+# - in_glossary.json: list of acronyms that are in the glossary markdown file mcr_meeting/app/infrastructure/llm/prompts/data/glossary.md
 # - not_in_glossary.json: list of acronyms that are not in the glossary markdown file but that we want to evaluate because they are present in the dataset audios.
 # - acronyms_not_evaluated.json: list of acronyms that are in the glossary but that we don't want to evaluate because they are too ambiguous or not relevant for the evaluation (e.g. "CAS" which is also a common French word).
 GLOSSARY_PATH = DATA_DIR / "in_glossary.json"
